% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ComputeWeight}
\alias{ComputeWeight}
\title{Compute the weight for each gene to construct gene-gene correlation}
\usage{
ComputeWeight(counts, normalization = TRUE, lib_size = NULL,
  method = "pearson")
}
\arguments{
\item{counts}{Raw count data, each column is a cell and each row is a gene.}

\item{normalization}{Bool variable to denote whether to do normalization before computing the weight, The default is TRUE.}

\item{lib_size}{Read deepth for each cell, For default, summarize all read counts across whole genes.}

\item{method}{The method to compute correlation matrix, pearson (default).}
}
\value{
Return the weight for each gene, p x 1 vector
}
\description{
Compute the weight for each gene to construct gene-gene correlation
}
