% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateiDEAObject}
\alias{CreateiDEAObject}
\title{Setup the iDEA object,}
\usage{
CreateiDEAObject(
  summary,
  annotation,
  project = "iDEA",
  max_var_beta = 100,
  min_precent_annot = 0.0025,
  num_core = 10
)
}
\arguments{
\item{summary}{Summary statistics (the estimated gene effect size and its variance, g1 x 2 matrix) from single-cell RNAseq differential expression tools, i.g. zingeR, MAST, etc. The summary statistics file should be in data.frame data format with gene name as row names, while the column names are not required but the order of the column matters: the first column should be the coefficient and the second column should be the variance of the coefficient for each gene.}

\item{annotation}{Pre-definted gene specific annotations (gene sets, g2 x m matrix), i.e., GO term, pathways etc. The gene specific annotation file is required data.frame data format with gene name as row names, while the header is the gene specific annotation name. The row names of annotation file should match exactly with row names of summary statistics and should be in the same type, i.e. gene symbol or transcription id etc. If not, one solution is to use biomaRt R package to convert the gene name to make sure they are consistent each other.}

\item{project}{Project name (string).}

\item{max_var_beta}{Include genes where the variances which are smaller than 'max_var_beta' are maintained.}

\item{min_precent_annot}{The threshold of coverage rate (CR), i.e., the number of annotated genes (gene set size) divided by the number of tested genes.}

\item{num_core}{Number of cores for parallel implementation.}
}
\value{
Returns a iDEA object with the summary statistics stored in object@summary and object@annotation.
object@project, object@num_core, object@num_gene, object@num_annot, object@gene_id, object@annot_id, are also initialized.
}
\description{
Setup the iDEA object,
}
