% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMMCMCStepSummary}
\alias{EMMCMCStepSummary}
\title{EM-MCMC function based on summary statistics, zscores}
\usage{
EMMCMCStepSummary(
  yin,
  varbetain,
  Ain,
  betain,
  tauin,
  em_iterin,
  mcmc_iterin,
  min_degenein
)
}
\arguments{
\item{yin}{Estimated beta}

\item{varbetain}{Variance}

\item{Ain}{Annotation}

\item{betain}{Initial beta}

\item{tauin}{Initial annotation coefficient}

\item{em_iterin}{EM iteration}

\item{mcmc_iterin}{MCMC iteration}

\item{min_degenein}{Minimum DE genes}
}
\value{
A list
}
\description{
EM-MCMC function based on summary statistics, zscores
}
