% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDEASummary.R
\name{GetSummaryData}
\alias{GetSummaryData}
\title{Get summary data from the existing DE method}
\usage{
GetSummaryData(counts, celltype, method = "DESeq2")
}
\arguments{
\item{counts}{scRNAseq raw counts matrix,  genes (row) by cells (column)}

\item{celltype}{cell types for counts matrix, two cell types are allowed}

\item{method}{DE method to obtain the summary data}
}
\value{
Returns summary data for iDEA
}
\description{
Get summary data from the existing DE method
}
