% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iDEA.FDR}
\alias{iDEA.FDR}
\title{We constructed a permuted null distribution by permuting gene labels. Specifically, we permuted the gene labels of annotation matrix to construced the null pvakue for gene sets. Then we calculate FDR for each gene sets given the permuted null distribution}
\usage{
iDEA.FDR(object.alt, object.null, numPermute = 10)
}
\arguments{
\item{object.alt}{iDEA object under the alternative. iDEA analysis in this object was performed using real gene sets}

\item{object.null}{iDEA object under the permuted null. iDEA analysis in this object was performed by permuting the gene labels in gene sets.}

\item{numPermute}{Number of permutation used in object.null. Default is 10.}
}
\value{
Returns a data frame which stores the results of iDEA with FDR values for each gene sets
}
\description{
We constructed a permuted null distribution by permuting gene labels. Specifically, we permuted the gene labels of annotation matrix to construced the null pvakue for gene sets. Then we calculate FDR for each gene sets given the permuted null distribution
}
