% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{iDEA}
\alias{iDEA}
\alias{iDEA-class}
\title{The iDEA Class}
\description{
The iDEA object stores all information
associated with the dataset, including data, annotations, analyes, etc.
}
\details{
Each iDEA object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{summary}}{The summary data from DE method}

\item{\code{annotation}}{The gene specific annotations}

\item{\code{project}}{The sub project name}

\item{\code{gene_id}}{The gene names}

\item{\code{annot_id}}{The annotation names}

\item{\code{weight}}{The weight for each gene for weighted version code}

\item{\code{num_gene}}{The number of genes}

\item{\code{num_core}}{The number of cores in parallel running}

\item{\code{null_model}}{The model under the null}

\item{\code{emmcmc}}{The results are from EM-MCMC step}

\item{\code{louis}}{The final results after Louis correction}
}}

