% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDEASummary.R
\name{iDEAWeight.fit}
\alias{iDEAWeight.fit}
\title{Integrative DE and GSEA based summary statistics and weight for each gene}
\usage{
iDEAWeight.fit(object, weight = NULL, fit_null = FALSE,
  init_beta = NULL, init_tau = c(-2, 0.5), min_degene = 5,
  em_iter = 20, mcmc_iter = 1000, fit.tol = 1e-05, verbose = TRUE,
  ...)
}
\arguments{
\item{object}{iDEA object}

\item{weight}{The weight for each gene to construct the gene-gene interation}

\item{fit_null}{Bool variable to indicate whether fitting the model without the annoation}

\item{init_beta}{Initial value for gene effect size, beta}

\item{init_tau}{Initial value for annotations, including the intercept}

\item{min_degene}{The threshold for the number of detected DE genes. For some of extremely cases, 
the method does not work when the number of detected DE genes is 0.}

\item{em_iter}{Maximum iteration for EM algorithm}

\item{mcmc_iter}{Maximum iteration for MCMC algorithm}

\item{fit.tol}{Tol for fitting the model}

\item{verbose}{Print the progresses}

\item{...}{Ignored}
}
\value{
Returns a iDEA object with EM-MCMC results stored in object@emmcmc.
}
\description{
Integrative DE and GSEA into a unfied framwork. The parameters of the model were infered via EM-MCMC algorithm.
}
