% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louismethod.R
\name{LouisCorrect}
\alias{LouisCorrect}
\title{Correct the p-values using Louis method}
\usage{
LouisCorrect(object)
}
\arguments{
\item{object}{iDEA object}

\item{...}{Ignored}
}
\value{
Returns a iDEA object with corrected p-values stored in object@louis.
}
\description{
A part of variance is missed when we use EM algorithm, since we should use the complete-data gradient vector or second derivative matrix rather than the incomplete data likelihood. 
See details: "Thomas A. Louis. Finding the Observed Information Matrix when Using the EM Algorithm, Journal of the Royal Statistical Society. 
Series B (Methodological) Vol. 44, No. 2 (1982), pp. 226-233".
}
