% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{QQNorm}
\alias{QQNorm}
\title{Quantile-Quantile normalization for count data}
\usage{
QQNorm(counts, lib_size = NULL)
}
\arguments{
\item{counts}{Raw count data, each column is a cell and each row is a gene.}

\item{lib_size}{Read deepth for each cell, For default, summarize all read counts across whole genes.}
}
\value{
Return normalized data
}
\description{
Quantile-Quantile normalization for count data
}
