% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IDEA.R
\name{runIDEA}
\alias{runIDEA}
\title{Run the default IDEA app for analysis locally}
\usage{
runIDEA()
}
\description{
\code{runIDEA} run IDEA locally
}
\author{
Qi Zhao
}
\seealso{
\code{IDEA} \link{http:\\\\idea.biocuckoo.org}
}

