% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateiDEAObject}
\alias{CreateiDEAObject}
\title{Setup the iDEA object,}
\usage{
CreateiDEAObject(summary = NULL, annotation, counts = NULL,
  cell_type = NULL, method = "DESeq2", project = "iDEA",
  max_var_beta = 100, min_precent_annot = 0.005, num_core = 10)
}
\arguments{
\item{summary}{Summary statistics (the estimated gene effect size and its variance, g1 x 2 matrix) from single-cell RNAseq differential expression tools, i.g. zingeR, MAST, etc.}

\item{annotation}{(required) Pre-definted gene specific annotations (gene sets, g2 x m matrix), i.e., GO term, pathways etc.}

\item{counts}{Raw scRNAseq counts to obtain the summary data}

\item{cell_type}{Cell type of the counts data}

\item{method}{the DE method to obtain the summary data}

\item{project}{Project name (string).}

\item{max_var_beta}{Include genes where the variances which are smaller than 'max_var_beta' are maintained  .}

\item{min_precent_annot}{The threshold of coverage rate (CR), i.e., the number of annotated genes (gene set size) divided by the number of tested genes.}

\item{num_core}{Number of cores for parallel implementation.}
}
\value{
Returns a iDEA object with the summary statistics stored in object@summary and object@annotation.
object@project, object@num_core, object@num_gene, object@num_annot, object@gene_id, object@annot_id, are also initialized.
}
\description{
Setup the iDEA object,
}
