% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDEASummary.R
\name{iDEA.BMA}
\alias{iDEA.BMA}
\title{Bayesian model averaging (BMA) approach to aggregate DE evidence for any given genes across all available gene sets without the requirement of pre-selecting a gene set.}
\usage{
iDEA.BMA(object)
}
\arguments{
\item{object}{iDEA object}

\item{...}{Ignored}
}
\value{
Returns a iDEA object with posterior inclusion probability averaged by all gene sets results in object@BMA_pip.
}
\description{
Bayesian model averaging (BMA) approach to aggregate DE evidence for any given genes across all available gene sets without the requirement of pre-selecting a gene set.
}
